requirements.txt
autopep8==1.5.6
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
Faker==8.1.0
fordev==1.0.3
idna==2.10
imageio==2.9.0
imutils==0.5.4
kiwisolver==1.3.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-python==4.1.2.30
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
tesseract==0.1.3
text-unidecode==1.3
tifffile==2021.4.8
toml==0.10.2
urllib3==1.26.4
